Introduce kube-apiserver proxy
Adds an local proxy to integration tests running by default and giving
access the the kube-apiserver and it's proxy capabilities.

It's used in the metrics tests to access the addon-operator-metrics
service on port :8080.

The proxy is environment independend and should work the same in local
development and in CI/CD environments.

Signed-off-by: Nico Schieder <code@nico-schieder.de>