Leader Election to use ADDON_OPERATOR_NAMESPACE
The LeaderElectionNamespace, when not set, defaults to
getInClusterNamespace(). This turns to be a problem when we are running
the manager off-cluster (e.g. Hypershift).

This patch sets the LeaderElectionNamespace explicitly to the value of
addonOperatorNamespace, which can be set via the
ADDON_OPERATOR_NAMESPACE env var and defaults to the in-cluster
Namespace.

Signed-off-by: Amador Pahim <apahim@redhat.com>