Expose pprof on localhost by default
Accessing pprof profiles for Addon Operator involved patching the
deployment or CSV, which we don't always have permission to do and
requires at the very least a pod restart.

This change exposes pprof on localhost by default to not expose it
openly to the whole cluster. Data can still be accessed by running curl
within the container to fetch pprof profiles.

Signed-off-by: Nico Schieder <nschieder@redhat.com>