Disable for cache for CSV resource
Add `OnlyMetadata` option to ClusterServiceVersion watch setup

Replace the use of `operatorsv1alpha1.ClusterServiceVersion` type
with `unstructured.Unstructured` in the `get` CSV operation. This
ensures that Get operations on CSV resource are not cached as
unstructured.Unstructured based client operations are not cached
by default in the manager default client.

Ref: https://github.com/kubernetes-sigs/controller-runtime/blob/cd0058ad295c268da1e7233e609a9a18dd60b5f6/pkg/client/split.go#L115

Signed-off-by: Nikhil Thomas <nikthoma@redhat.com>